lesson6/l6-llm-api-client/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
aiohttp==3.9.3
python-dotenv==1.0.1
google-generativeai==0.4.0
redis==5.0.1
prometheus-client==0.19.0
structlog==24.1.0
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
